Updated PC to accept Writeback
Also made it be able to auto increment by 2, have a Enable input that will stop the auto increment. It has 2 other new inputs to do with the writback. Qin which is the address value from the writeback and then br which sets the program counter latch to be the writeback value. I have added it to the processor but have not connected up the new inputs to any control.