Fixed the fetch stage of the pipline.
Added a program counter that is ready to receive a write back. Added nop insertion, however I'm not if the program counter is behaving correctly when nop_insertion is happening. Apart from that the fetch stage of the pipline seems to be working fine. Have also implemented the control after the fetch stage but I haven't connected it up yet.