Added nop insertion
The PC value and instruction go getting propagated to the next stage. One thing I'm unsure of is whether we want to propagate the PC for the current instruction or the incremented one. Currently I'm propagating the incremented PC. 

I've also started implementing the control logic for the decoding stage.